requirements.txt
Flask==2.3.3
pymongo==4.6.0
python-dotenv==1.0.0
Werkzeug==2.3.3
